fix: cors is disabled for auth and client endpoints
QW-126
