+++ app/controllers/client_controller.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, request, jsonify, make_response
 
+from app.controllers import _build_cors_preflight_response, _corsify_actual_response
 from app.schemas.client_register_dto import ClientRegisterDTO
 from app.usecases.register_client import RegisterClientUseCase
 
@@ -19,16 +20,3 @@ def register_client():
             return jsonify({"error": str(e)}), 409
         except Exception as e:
             return jsonify({"error": str(e)}), 400
-
-
-def _build_cors_preflight_response():
-    response = make_response()
-    response.headers.add("Access-Control-Allow-Origin", "*")
-    response.headers.add('Access-Control-Allow-Headers', "*")
-    response.headers.add('Access-Control-Allow-Methods', "*")
-    return response
-
-
-def _corsify_actual_response(response):
-    response.headers.add("Access-Control-Allow-Origin", "*")
-    return response
